+++ lib/employee/employee_profile_page.dart
@@ -303,9 +303,9 @@ class _EmployeeProfilPageState extends State<EmployeeProfilPage> {
                         ),
                         color: GREEN,
                         onPressed: () {
-                          int hours;
+                          double hours;
                           try {
-                            hours = int.parse(_hoursController.text);
+                            hours = double.parse(_hoursController.text);
                           } catch (FormatException) {
                             ToastService.showErrorToast(getTranslated(context, 'givenValueIsNotANumber'));
                             return;
